+++ src/index.ts
@@ -1,7 +1,7 @@
 export { TOP_LEVEL, DEFAULT_LEVELS as defaultLevels } from "./constants.js";
 export { createLog } from "./core/create_log.js";
 export { normalizeLevels } from "./levels/index.js";
-export { getEntriesForDir } from "./storage/query.js";
+export { getLogsForDir } from "./storage/query.js";
 export { logStream } from "./stream/index.js";
 
 export type {
@@ -13,6 +13,9 @@ export type {
   LogOrigin,
   LogQueryOptions,
   LogQueryResult,
+  LogQueryTotals,
+  LogPartitionTotals,
+  LogPartitionSummary,
   LogStats,
   LogStreamName,
   RedactOptions,